fix(cc-components): rename out-dial to outdial
