+++ packages/contact-center/cc-components/src/components/task/OutdialCall/outdial-call.styles.scss
