+++ packages/contact-center/cc-components/src/components/task/OutdialCall/outdial-call.tsx
@@ -1,6 +1,6 @@
 import React, {useState} from 'react';
 import {OutdialCallComponentProps} from '../task.types';
-import './out-dial-call.styles.scss';
+import './outdial-call.style.scss';
 
 const OutdialCallComponent: React.FunctionComponent<OutdialCallComponentProps> = (props) => {
   const {startOutdial} = props;
